+++ src/paint_area.hpp
